Make IfcBoolean and IfcLogic return the actual value.
In previous implementations IfcBoolean and IfcLogical returned "T" for True, "F" for False or "U" for Unknown  – in the case of IfcLogical –. With the latest updates regardless of the actual value when calling GetLine IfcBoolean and IfcLogical get always false. This may be solved by checking for checking for the string "T" instead of "true", but the proposal here is to leave the value defined in the ifc file, otherwise with a real boolean value would be hard to identify IfcLogical's Unknown.