Preserve null values when parsing number, boolean and logical.
When parsing IFC numbers, boolean and logical, null values are not preserved. Numbers – e.g. IFCREAL or IFCINTEGER – are converted to NaN, booleans – IFCBOOLEAN – to false and logicals – IFCLOGICAL – to UNKNOWN. Since these conversions may be subjective, it would be more transparent to leave the API consumer to handle it.